requirements.txt
PyQt6
psycopg2-binary
PyYAML
keyring
